Create update-teamviewer
This script automatically checks the supported filetype of the Linux version (RPM or DEB), and checks the version of the installed TeamViewer app installed currently against the latest version available on teamviewer.com.  Sudo access is required to be able to install the update, but not to run the script.